Visual tests: code changes
