+++ lib/iris/tests/test_quickplot.py
@@ -1,4 +1,4 @@
-# (C) British Crown Copyright 2010 - 2015, Met Office
+# (C) British Crown Copyright 2016, Met Office
 #
 # This file is part of Iris.
 #
@@ -14,103 +14,89 @@
 #
 # You should have received a copy of the GNU Lesser General Public License
 # along with Iris.  If not, see <http://www.gnu.org/licenses/>.
-"""
-Tests the high-level plotting interface.
-
-"""
+"""Integration tests for the package :mod:`iris.quickplot`."""
 
 from __future__ import (absolute_import, division, print_function)
 from six.moves import (filter, input, map, range, zip)  # noqa
+import six
 
-# import iris tests first so that some things can be initialised before importing anything else
+# import iris tests first so that some things can be initialised before
+# importing anything else
 import iris.tests as tests
-import iris.tests.test_plot as test_plot
-
 import iris
+from iris.tests.integration.plotting import \
+     TestNDCoordinatesGiven, LambdaStr, load_cube_once, load_theta
 
 # Run tests in no graphics mode if matplotlib is not available.
 if tests.MPL_AVAILABLE:
     import matplotlib.pyplot as plt
-    import iris.plot as iplt
-    import iris.quickplot as qplt
-
-
-# Caches _load_theta so subsequent calls are faster
-def cache(fn, cache={}):
-    def inner(*args, **kwargs):
-        key = "result"
-        if not cache:
-            cache[key] =  fn(*args, **kwargs)
-        return cache[key]
-    return inner
-
-
-@cache
-def _load_theta():
-    path = tests.get_data_path(('PP', 'COLPEX', 'theta_and_orog_subset.pp'))
-    theta = iris.load_cube(path, 'air_potential_temperature')
-
-    # Improve the unit
-    theta.units = 'K'
-
-    return theta
+    from iris.plot import (plot, contour, contourf, pcolor, pcolormesh, points)
 
 
 @tests.skip_data
 @tests.skip_plot
-class TestQuickplotCoordinatesGiven(test_plot.TestPlotCoordinatesGiven):
+class TestAttributePositive(tests.GraphicsTest):
+    def test_1d_positive_up(self):
+        path = tests.get_data_path(('NetCDF', 'ORCA2', 'votemper.nc'))
+        cube = iris.load_cube(path)
+        plot(cube.coord('depth'), cube[0, :, 60, 80])
+        self.check_graphic()
+
+    def test_1d_positive_down(self):
+        path = tests.get_data_path(('NetCDF', 'ORCA2', 'votemper.nc'))
+        cube = iris.load_cube(path)
+        plot(cube[0, :, 60, 80], cube.coord('depth'))
+        self.check_graphic()
+
+    def test_2d_positive_up(self):
+        path = tests.get_data_path(('NetCDF', 'testing',
+                                    'small_theta_colpex.nc'))
+        cube = iris.load_cube(path)[0, :, 42, :]
+        pcolormesh(cube)
+        self.check_graphic()
+
+    def test_2d_positive_down(self):
+        path = tests.get_data_path(('NetCDF', 'ORCA2', 'votemper.nc'))
+        cube = iris.load_cube(path)[0, :, 42, :]
+        pcolormesh(cube)
+        self.check_graphic()
+
+
+@tests.skip_data
+@tests.skip_plot
+class TestQuickplotCoordinatesGiven(TestNDCoordinatesGiven):
     def setUp(self):
-        filename = tests.get_data_path(('PP', 'COLPEX', 'theta_and_orog_subset.pp'))
-        self.cube = test_plot.load_cube_once(filename, 'air_potential_temperature')
+        filename = tests.get_data_path(
+            ('PP', 'COLPEX', 'theta_and_orog_subset.pp'))
+        self.cube = load_cube_once(filename, 'air_potential_temperature')
 
-        self.draw_module = iris.quickplot
-        self.contourf = test_plot.LambdaStr('iris.quickplot.contourf', lambda cube, *args, **kwargs:
-                                                  iris.quickplot.contourf(cube, *args, **kwargs))
-        self.contour = test_plot.LambdaStr('iris.quickplot.contour', lambda cube, *args, **kwargs:
-                                                  iris.quickplot.contour(cube, *args, **kwargs))
-        self.points = test_plot.LambdaStr('iris.quickplot.points', lambda cube, *args, **kwargs:
-                                                  iris.quickplot.points(cube, c=cube.data, *args, **kwargs))
-        self.plot = test_plot.LambdaStr('iris.quickplot.plot', lambda cube, *args, **kwargs:
-                                                  iris.quickplot.plot(cube, *args, **kwargs))
+        self.contourf = LambdaStr('iris.quickplot.contourf',
+                                  lambda cube, *args, **kwargs:
+                                  contourf(cube, *args, **kwargs))
+        self.contour = LambdaStr('iris.quickplot.contour',
+                                 lambda cube, *args, **kwargs:
+                                 contour(cube, *args, **kwargs))
+        self.pcolor = LambdaStr('iris.quickplot.pcolor',
+                                lambda cube, *args, **kwargs:
+                                pcolor(cube, *args, **kwargs))
+        self.pcolormesh = LambdaStr('iris.quickplot.pcolormesh',
+                                    lambda cube, *args, **kwargs:
+                                    pcolormesh(cube, *args, **kwargs))
+        self.points = LambdaStr('iris.quickplot.points',
+                                lambda cube, *args, **kwargs:
+                                points(cube, c=cube.data, *args, **kwargs))
+        self.plot = LambdaStr('iris.quickplot.plot',
+                              lambda cube, *args, **kwargs:
+                              plot(cube, *args, **kwargs))
 
-        self.results = {'yx': (
-                           [self.contourf, ['grid_latitude', 'grid_longitude']],
-                           [self.contourf, ['grid_longitude', 'grid_latitude']],
-                           [self.contour, ['grid_latitude', 'grid_longitude']],
-                           [self.contour, ['grid_longitude', 'grid_latitude']],
-                           [self.points, ['grid_latitude', 'grid_longitude']],
-                           [self.points, ['grid_longitude', 'grid_latitude']],
-                           ),
-                       'zx': (
-                           [self.contourf, ['model_level_number', 'grid_longitude']],
-                           [self.contourf, ['grid_longitude', 'model_level_number']],
-                           [self.contour, ['model_level_number', 'grid_longitude']],
-                           [self.contour, ['grid_longitude', 'model_level_number']],
-                           [self.points, ['model_level_number', 'grid_longitude']],
-                           [self.points, ['grid_longitude', 'model_level_number']],
-                           ),
-                        'tx': (
-                           [self.contourf, ['time', 'grid_longitude']],
-                           [self.contourf, ['grid_longitude', 'time']],
-                           [self.contour, ['time', 'grid_longitude']],
-                           [self.contour, ['grid_longitude', 'time']],
-                           [self.points, ['time', 'grid_longitude']],
-                           [self.points, ['grid_longitude', 'time']],
-                           ),
-                        'x': (
-                           [self.plot, ['grid_longitude']],
-                           ),
-                        'y': (
-                           [self.plot, ['grid_latitude']],
-                           ),
-                       }
+        super(TestQuickplotCoordinatesGiven, self).__init__()
 
 
 @tests.skip_data
 @tests.skip_plot
 class TestLabels(tests.GraphicsTest):
     def setUp(self):
-        self.theta = _load_theta()
+        self.theta = load_theta()
 
     def _slice(self, coords):
         """Returns the first cube containing the requested coordinates."""
@@ -119,65 +105,62 @@ class TestLabels(tests.GraphicsTest):
         return cube
 
     def _small(self):
-        # Use a restricted size so we can make out the detail
+        # Use a restricted size so we can make out the detail.
         cube = self._slice(['model_level_number', 'grid_longitude'])
         return cube[:5, :5]
 
     def test_contour(self):
-        qplt.contour(self._small())
+        contour(self._small())
         self.check_graphic()
 
-        qplt.contourf(self._small(), coords=['model_level_number', 'grid_longitude'])
+        contour(self._small(),
+                coords=['model_level_number', 'grid_longitude'])
         self.check_graphic()
 
     def test_contourf(self):
-        qplt.contourf(self._small())
-
-        cube = self._small()
-        iplt.orography_at_points(cube)
-
+        contourf(self._small())
         self.check_graphic()
 
-        qplt.contourf(self._small(), coords=['model_level_number', 'grid_longitude'])
+        contourf(self._small(),
+                 coords=['model_level_number', 'grid_longitude'])
         self.check_graphic()
 
-        qplt.contourf(self._small(), coords=['grid_longitude', 'model_level_number'])
+        contourf(self._small(),
+                 coords=['grid_longitude', 'model_level_number'])
         self.check_graphic()
 
     def test_contourf_nameless(self):
         cube = self._small()
         cube.standard_name = None
-        qplt.contourf(cube, coords=['grid_longitude', 'model_level_number'])
+        contourf(cube,
+                 coords=['grid_longitude', 'model_level_number'])
         self.check_graphic()
 
     def test_pcolor(self):
-        qplt.pcolor(self._small())
+        pcolor(self._small())
         self.check_graphic()
 
     def test_pcolormesh(self):
-        qplt.pcolormesh(self._small())
-
-        #cube = self._small()
-        #iplt.orography_at_bounds(cube)
-
+        pcolormesh(self._small())
         self.check_graphic()
 
     def test_map(self):
         cube = self._slice(['grid_latitude', 'grid_longitude'])
-        qplt.contour(cube)
+        contour(cube)
         self.check_graphic()
 
-        # check that the result of adding 360 to the data is *almost* identically the same result
+    def test_add_roll(self):
+        # Check that the result of adding 360 to the data is almost identical.
+        cube = self._slice(['grid_latitude', 'grid_longitude'])
         lon = cube.coord('grid_longitude')
         lon.points = lon.points + 360
-        qplt.contour(cube)
+        contour(cube)
         self.check_graphic()
 
     def test_alignment(self):
         cube = self._small()
-        qplt.contourf(cube)
-        #qplt.outline(cube)
-        qplt.points(cube)
+        contourf(cube)
+        points(cube)
         self.check_graphic()
 
 
@@ -191,13 +174,13 @@ class TestTimeReferenceUnitsLabels(tests.GraphicsTest):
 
     def test_reference_time_units(self):
         # units should not be displayed for a reference time
-        qplt.plot(self.cube.coord('time'), self.cube)
+        plot(self.cube.coord('time'), self.cube)
         plt.gcf().autofmt_xdate()
         self.check_graphic()
 
     def test_not_reference_time_units(self):
         # units should be displayed for other time coordinates
-        qplt.plot(self.cube.coord('forecast_period'), self.cube)
+        plot(self.cube.coord('forecast_period'), self.cube)
         self.check_graphic()
